+++ IdentityServer4.NHibernate.IntegrationTests/TestStorage/TestDatabaseBuilder.cs
@@ -1,9 +1,11 @@
-﻿using System.Diagnostics;
+﻿using System;
+using System.Diagnostics;
+using System.IO;
 using IdentityServer4.NHibernate.Database;
 using IdentityServer4.NHibernate.Extensions;
 using IdentityServer4.NHibernate.Options;
-using NHibernate.Cfg;
 using NHibernate.Tool.hbm2ddl;
+using Environment = NHibernate.Cfg.Environment;
 
 namespace IdentityServer4.NHibernate.IntegrationTests.TestStorage
 {
@@ -12,28 +14,27 @@ namespace IdentityServer4.NHibernate.IntegrationTests.TestStorage
     /// </summary>
     internal static class TestDatabaseBuilder
     {
-        public static SQLServerTestDatabase SQLServer2012TestDatabase(string serverName, string databaseName, ConfigurationStoreOptions configurationStoreOptions, OperationalStoreOptions operationalStoreOptions)
+        internal static SQLServerTestDatabase SQLServer2012TestDatabase(string connectionString, string databaseName, ConfigurationStoreOptions configurationStoreOptions, OperationalStoreOptions operationalStoreOptions)
         {
-            var connString = $"Data Source={serverName}; Initial Catalog={databaseName}; Integrated Security=SSPI; Application Name=IdentityServer4.NHibernate.IntegrationTests";
+            var dbConfig = Databases.SqlServer2012()
+                .UsingConnectionString($"{connectionString};Initial Catalog={databaseName}")
+                .EnableSqlStatementsLogging()
+                .AddConfigurationStoreMappings(configurationStoreOptions)
+                .AddOperationalStoreMappings(operationalStoreOptions)
+                .SetProperty(Environment.Hbm2ddlAuto, "create-drop");
+            var testDb = new SQLServerTestDatabase(connectionString, databaseName, dbConfig);
 
-            SQLServerTestDatabase testDb = null;
             try
             {
-                var dbConfig = Databases.SqlServer2012()
-                    .UsingConnectionString(connString)
-                    .EnableSqlStatementsLogging()
-                    .AddConfigurationStoreMappings(configurationStoreOptions)
-                    .AddOperationalStoreMappings(operationalStoreOptions)
-                    .SetProperty(Environment.Hbm2ddlAuto, "create-drop");
-
-                testDb = new SQLServerTestDatabase(serverName, databaseName, dbConfig);
+                testDb.DropIfExists();
                 testDb.Create();
                 new SchemaExport(dbConfig).Execute(false, true, false);
             }
-            catch (System.Exception ex)
+            catch (Exception ex)
             {
                 Debug.WriteLine(ex.Message);
-                testDb?.Drop();
+                testDb.DropIfExists();
+                throw;
             }
 
             return testDb;
@@ -42,25 +43,25 @@ namespace IdentityServer4.NHibernate.IntegrationTests.TestStorage
         internal static SQLiteTestDatabase SQLiteTestDatabase(string fileName, ConfigurationStoreOptions configurationStoreOptions, OperationalStoreOptions operationalStoreOptions)
         {
             var connString = $"Data Source={fileName};Version=3;Pooling=True;Synchronous=Full;";
+            var dbConfig = Databases.SQLite()
+                .UsingConnectionString(connString)
+                .EnableSqlStatementsLogging()
+                .AddConfigurationStoreMappings(configurationStoreOptions)
+                .AddOperationalStoreMappings(operationalStoreOptions)
+                .SetProperty(Environment.Hbm2ddlAuto, "create-drop");
+            var testDb = new SQLiteTestDatabase(fileName, dbConfig);
 
-            SQLiteTestDatabase testDb = null;
             try
             {
-                var dbConfig = Databases.SQLite()
-                    .UsingConnectionString(connString)
-                    .EnableSqlStatementsLogging()
-                    .AddConfigurationStoreMappings(configurationStoreOptions)
-                    .AddOperationalStoreMappings(operationalStoreOptions)
-                    .SetProperty(Environment.Hbm2ddlAuto, "create-drop");
-
-                testDb = new SQLiteTestDatabase(fileName, dbConfig);
+                testDb.DropIfExists();
                 testDb.Create();
                 new SchemaExport(dbConfig).Execute(false, true, false);
             }
-            catch (System.Exception ex)
+            catch (Exception ex)
             {
                 Debug.WriteLine(ex.Message);
-                testDb.Drop();
+                testDb.DropIfExists();
+                throw;
             }
 
             return testDb;
@@ -68,21 +69,73 @@ namespace IdentityServer4.NHibernate.IntegrationTests.TestStorage
 
         internal static SQLiteInMemoryTestDatabase SQLiteInMemoryTestDatabase(ConfigurationStoreOptions configurationStoreOptions, OperationalStoreOptions operationalStoreOptions)
         {
-            SQLiteInMemoryTestDatabase testDb = null;
+            var dbConfig = Databases.SQLiteInMemory()
+                .EnableSqlStatementsLogging()
+                .AddConfigurationStoreMappings(configurationStoreOptions)
+                .AddOperationalStoreMappings(operationalStoreOptions);
+            var testDb = new SQLiteInMemoryTestDatabase(dbConfig);
+
             try
             {
-                var dbConfig = Databases.SQLiteInMemory()
-                    .EnableSqlStatementsLogging()
-                    .AddConfigurationStoreMappings(configurationStoreOptions)
-                    .AddOperationalStoreMappings(operationalStoreOptions);
-
-                testDb = new SQLiteInMemoryTestDatabase(dbConfig);
                 testDb.Create();
-                new SchemaExport(dbConfig).Execute(false, true, false, testDb.ActiveConnection, null);
+                new SchemaExport(dbConfig).Execute(false, true, false, testDb.ActiveConnection, TextWriter.Null);
             }
-            catch (System.Exception ex)
+            catch (Exception ex)
             {
                 Debug.WriteLine(ex.Message);
+                throw;
+            }
+
+            return testDb;
+        }
+
+        internal static PostgreSQLTestDatabase PostgreSQLTestDatabase(string connectionString, string databaseName, ConfigurationStoreOptions configurationStoreOptions, OperationalStoreOptions operationalStoreOptions)
+        {
+            var dbConfig = Databases.PostgreSQL()
+                .UsingConnectionString($"{connectionString};Database={databaseName}")
+                .EnableSqlStatementsLogging()
+                .AddConfigurationStoreMappings(configurationStoreOptions)
+                .AddOperationalStoreMappings(operationalStoreOptions)
+                .SetProperty(Environment.Hbm2ddlAuto, "create-drop");
+            var testDb = new PostgreSQLTestDatabase(connectionString, databaseName, dbConfig);
+            
+            try
+            {
+                testDb.DropIfExists();
+                testDb.Create();
+                new SchemaExport(dbConfig).Execute(false, true, false);
+            }
+            catch (Exception ex)
+            {
+                Debug.WriteLine(ex.Message);
+                testDb.DropIfExists();
+                throw;
+            }
+
+            return testDb;
+        }
+
+        internal static MySqlTestDatabase MySQLTestDatabase(string connectionString, string databaseName, ConfigurationStoreOptions configurationStoreOptions, OperationalStoreOptions operationalStoreOptions)
+        {
+            var dbConfig = Databases.MySql()
+                .UsingConnectionString($"{connectionString};Database={databaseName}")
+                .EnableSqlStatementsLogging()
+                .AddConfigurationStoreMappings(configurationStoreOptions)
+                .AddOperationalStoreMappings(operationalStoreOptions)
+                .SetProperty(Environment.Hbm2ddlAuto, "create-drop");
+            var testDb = new MySqlTestDatabase(connectionString, databaseName, dbConfig);
+
+            try
+            {
+                testDb.DropIfExists();
+                testDb.Create();
+                new SchemaExport(dbConfig).Execute(false, true, false);
+            }
+            catch (Exception ex)
+            {
+                Debug.WriteLine(ex.Message);
+                testDb.DropIfExists();
+                throw;
             }
 
             return testDb;